requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
matplotlib>=3.7.0
tkcalendar>=1.6.1
